+++ tests/functional/test__ex_substitute.py
@@ -45,6 +45,20 @@ class Test_ex_substitute(unittest.FunctionalTestCase):
         self.eq('axxa\n|bxXb\ncxxc\n', ':substitute/x/y/gi', 'axxa\n|byyb\ncxxc\n')
         self.eq('axxa\n|bXxb\ncxxc\n', ':substitute/x/y/gi', 'axxa\n|byyb\ncxxc\n')
 
+    def test_i_flag_ignore_case(self):
+        self.set_option('ignorecase', True)
+        self.eq('|aA', ':s/a/x/g', '|xx')
+        self.eq('|aA', ':s/a/x/gi', '|xx')
+        self.set_option('ignorecase', False)
+        self.eq('|aA', ':s/a/x/g', '|xA')
+        self.eq('|aA', ':s/a/x/gi', '|xx')
+
+    def test_I_flag_dont_ignore_case(self):
+        self.set_option('ignorecase', True)
+        self.eq('|aA', ':s/a/x/gI', '|xA')
+        self.set_option('ignorecase', False)
+        self.eq('|aA', ':s/a/x/gI', '|xA')
+
     def test_ranges(self):
         self.eq('axxa\n|bxxb\ncxxc\n', ':1,$substitute/x/y/', 'ayxa\nbyxb\n|cyxc\n')
         self.eq('axxa\n|bxxb\ncxxc\n', ':1,$substitute/x/y/g', 'ayya\nbyyb\n|cyyc\n')